Switch from zsh to bash to bootstrap fish shell.
zsh has a weird problem on lxplus where it hangs when it's started in a
screen/tmux session. After a ctrl+c, executed after some seemingly
arbitrary number of seconds, the sheel carries on as normal and fish is
loaded.
bash doesn't have this problem, and seeing as I don't use zsh for
anything I might as well use bash as the default shell (which only
starts fish anyway).

This makes things like starting tmux on login by default is more
practical.