Clean up fish shell environment configuration.
Prefer to use `set -U` rather than setting things in files. It’s nicer
because the values were defined at runtime but never changed, so
better to use `set -U` which saves the hard-coded paths to a file.

I’ve left the `set` commands I use in the files, for easier setup of
new systems.