Add environment refresher shell function.
Need to do this because we can start and reattach tmux sessions across
different sway instances (e.g. if we close the window manager without
closing a tmux session, and then start a new sway instance and reattach
to the same tmux session). The new function makes sure the sway-related
environment variables are updated before any command is run.